Rename tests/util to tests/nats_server
